--- index2.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <title>Урок "Модули"</title>
-  </head>
-
-  <body>
-    <h1>Список задач</h1>
-    <ul class="tasks" id="list"></ul>
-    <br />
-    <!— Должна появиться после входа —>
-    <div class="form">
-      <h3 class="form-title">Форма добавления</h3>
-      <p class="form-name">Имя пользователя:</p>
-      <div class="form-row">
-        Что нужно сделать:
-        <input
-          type="text"
-          id="text-input"
-          class="input"
-          placeholder="Выпить кофе"
-        />
-      </div>
-      <br />
-      <button class="button" id="add-button">обавить</button>
-    </div>
-  </body>
-  <script type="module" src="main2.js">
-    
-  </script>
-</html>
